Change copy about pure functions
I wasn't 100% sure what was the intention here: to say that date-fns is
built with pure functions or that the user will build their app with
pure functions from date-fns.

This change assumes that the intention was to say that date-fns exposes
pure functions and that this is what the user will be building with.